requirements.txt
rpi-rgb-led-matrix==1.0.0
Pillow==10.0.0
pytz==2023.3
requests==2.31.0
timezonefinder==6.2.0
geopy==2.4.1 